backend/requirements.txt
fastapi
uvicorn
google-generativeai>=0.7.0
google-genai
pydantic
pydantic-settings
python-dotenv
httpx
firebase-admin
gunicorn
